+++ SeeD/src/UI.h
@@ -905,7 +905,7 @@ public:
             return;
         }
 
-        uint width = ImGui::GetContentRegionAvail().x * 0.5f - ImGui::GetStyle().ItemSpacing.x * 0.5f;
+        float width = ImGui::GetContentRegionAvail().x * 0.5f - ImGui::GetStyle().ItemSpacing.x * 0.5f;
 
         if (ImGui::BeginTable("cpuTable", 6, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg | ImGuiTableFlags_ScrollY | ImGuiTableFlags_HighlightHoveredColumn, ImVec2(width, 0)))
         {
